Adds triggered event "clear.datatable"
When the injected "Clear" link is clicked, it now emits a ``clear.datatable`` event, which can be detected higher on the DOM by any external script.  Like any handler, the event itself is the first argument, and a second argument is a reference to the initialized datatable (oTable) object itself.