Remove ambiguous handling of improper calls to keyed helpers
A "keyed" helper is a helper that is decorated with a wrapping function that checks to see if a "key" kwarg is sent (returns another wrapper over the real helper to do the "key" conversion at runtime) or else an "instance" argument, which implies the real helper is being used directly (either in a custom callback or as a raw reference in the datatable options callback designation).

If neither was present, it means that user called the helper without any arguments, which is effectively an invalid interaction.  Previously, there was logic to swallow the unusable interaction and return the helper again, but this isn't intuitive or obvious from outside the API.

A ValueError will now be raised informing of the unsupported interaction.