Rejigger return value of apply_queryset_options()
This method was returning a 3-tuple of (object_list, length,
filtered_length), which was required information, but was
counter-intuitive when serialize_to_json() was having to unpack its
single argument into three separate ones.

This solution preserves the single argument format, which is a list
subclass with data attributes on it.  This approach was chosen to
prevent the renaming of the get_object_list() method to something less
comprehensible.