Add external dictionary of model field classes
This should allow external customization of how field classes are
treated, particularly for external field from third-party apps, and
fields that we need to dynamically detect, such as the overlooked
``GenericIPAddressField``.

This closes #53 and address #49, although in the latter case, we may
still eventually want a way to register a field type “category” into
the dictionary with a set of trailing ORM query bits that can be
applied to support unknown extensions to the query language.