Rewrite of how to declare DTV options
Options were previous contained within a dictionary on the class, but
this was a problem in a couple of ways: modifying the declared options
per-request was a pain in the head, and nesting all options goes
against the established pattern of top-level class attributes for view
configuration.

This new strategy is to follow more closely what the django.forms
framework does: declare a Datatable object that models how the table
will work.  A Datatable can declare a Meta inner class to configure all
the options we supported in the old datatable_options view attribute.

A view now need only reference the Datatable class it wants to
leverage, and that's it.  All callbacks for values are on the new
options object.

Like the Django class-based form views, you can also declare the
options on the class as loose attributes.  These will override the
options on the datatatable class' Meta config.

Still missing some documentation fixes, and MultipleTablesDatatableView
example is busted up pretty bad until I rework how it functions for
this new code.