Add search/sort API to Datatable
We have a functional difference from the 0.8 series right now, in that
multiple terms are considered OR terms, not AND terms.  This will
eventually be corrected, but needs to have some of the looping reversed
so that terms are requested from the datatable level, and not embedded
inside the column's loop.