Add request_method option to Meta
Fixes #146
Default method is still GET, but by setting this to POST, the view will
now detect the connection between the datatable's desired request
method and the method being used.

The implementation of the "get" method on the view has been replaced
with "dispatch" in order to avoid duplicating the get() logic.  This
simplifies the usage pattern if the view already implements some other
post() behavior.