Add 'cache_queryset_count' option
Provides an option to cache the unfiltered queryset count with the same
caching context as the queryset itself.  Might be a niche option, but
it would be a pain to cook this up from scratch.

Note that no caching is in place for the filtered_objects list.  If you
want that (for example, for the limited utility of caching based on
search terms) you should override ``count_objects()`` and arrange for
that.