Remove Travis configuration, CHANGELOG, and requirements files
This commit completely removes the Travis configuration file (.travis.yml), the change log (CHANGELOG.md), and several requirements files (requirements.txt, requirements-docs.txt, and requirements_dev.txt). This is part of a wider effort to remove unnecessary or outdated files and clean up the repository. The deletion of these files results in a cleaner project structure and the reduction of future maintenance or updates for those files which are no longer needed.